KTOR-1144 Make OAuth2 functionality multiplatform (#3350)
