+++ ktor-server/ktor-server-plugins/ktor-server-auth/jvm/src/io/ktor/server/auth/OAuth1a.kt
@@ -15,6 +15,8 @@ import io.ktor.server.application.*
 import io.ktor.server.response.*
 import io.ktor.util.*
 import io.ktor.util.pipeline.*
+import io.ktor.utils.io.errors.*
+import io.ktor.utils.io.errors.IOException
 import kotlinx.coroutines.*
 import java.io.*
 import java.time.*
